Reorganize

+++ src/Integer.ts
@@ -1,19 +1,18 @@
-import { newError } from '../source/newError'
 import { Brand } from './Brand'
 
 // Error is the biggest value
-export type Float = Brand<number | Error, 'Float'>
+export type Integer = Brand<number | Error, 'Integer'>
 
-type E = Brand<Error, 'Float'>
+type E = Brand<Error, 'Integer'>
 
-export const additionOverflow = (a: Float, b: Float) => {
+export const additionOverflow = (a: Integer, b: Integer) => {
   const ADDITION_OVERFLOW = new Error(
     `Addition between ${a} and ${b} will overflow`
   )
   return ADDITION_OVERFLOW as E
 }
 
-export const multiplicationOverflow = (a: Float, b: Float) => {
+export const multiplicationOverflow = (a: Integer, b: Integer) => {
   const MULTIPLICATION_OVERFLOW = new Error(
     `Multiplication between ${a} and ${b} will overflow`
   )
